chore: remove ublue copr for ptyxis install (#271)
* Remove ublue copr for ptyxis install

* Only use COPR for 39

* Source /etc/os-release

+++ lumina/scripts/_base/003-ptyxis-terminal.sh
@@ -5,8 +5,15 @@ source /etc/os-release
 
 FEDORA_VERSION=$VERSION_ID
 
-wget https://copr.fedorainfracloud.org/coprs/ublue-os/staging/repo/fedora-$(rpm -E %fedora)/ublue-os-staging-fedora-$(rpm -E %fedora).repo -O /etc/yum.repos.d/_copr_ublue-os_staging.repo
+if [[ "$FEDORA_VERSION" == "39" ]]; then
+  # Add the COPR repository
+  wget https://copr.fedorainfracloud.org/coprs/ublue-os/staging/repo/fedora-$(rpm -E %fedora)/ublue-os-staging-fedora-$(rpm -E %fedora).repo -O /etc/yum.repos.d/_copr_ublue-os_staging.repo
+fi
 
+# Install the package
 rpm-ostree install ptyxis
 
-rm -rf /etc/yum.repos.d/_copr_ublue-os_staging.repo
+# Check if the Fedora version is 39
+if [[ "$FEDORA_VERSION" == "39" ]]; then
+  rm -rf /etc/yum.repos.d/_copr_ublue-os_staging.repo
+fi
